p8fs-node/pyproject.toml
[project]
name = "p8fs-node"
version = "1.1.27"
description = "Core processing unit for content handling and embeddings"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.15",
    "p8fs-cluster",
    "pydantic>=2.5.0",
    "nats-py>=2.6.0",
    "aioboto3>=12.0.0",
    "kreuzberg>=0.1.0",
    "pypdf>=3.17.0",
    "python-docx>=1.1.0",
    "docx2txt>=0.8",
    "odfpy>=1.4.1",
    "striprtf>=0.0.26",
    "markdown>=3.5.0",
    "pillow>=11.0.0",
    "numpy>=1.26.0",
    "httpx>=0.25.0",
    "urllib3>=2.2.3",
    "certifi>=2024.7.4",
    "setuptools>=75.3.0",
    "pydub>=0.25.0",
    "chardet>=5.0.0",
    "pymupdf>=1.23.0",
    "pyyaml>=6.0",
    "typer>=0.9.0",
    "semchunk>=3.2.3",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
]

[project.scripts]
p8fs-node = "p8fs_node.cli.main:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.12.0",
]
audio = [
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.uv.sources]
p8fs-cluster = { workspace = true }

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.1.0",
    "pytest-mock>=3.15.0",
]
